demo按能力划分模块、录音转写demo更新
