+++ src/main/java/cn/xfyun/demo/face/AntiSpoofClientApp.java
@@ -1,48 +1,45 @@
-package cn.xfyun.demo;
+package cn.xfyun.demo.face;
 
-import cn.xfyun.api.TupApiClient;
-import cn.xfyun.config.PropertiesConfig;
-import cn.xfyun.config.TupApiEnum;
 import cn.hutool.core.io.IoUtil;
-
+import cn.xfyun.api.AntiSpoofClient;
+import cn.xfyun.config.PropertiesConfig;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.InputStream;
 import java.net.URISyntaxException;
+import java.util.Base64;
 
 /**
  * @author mqgao
  * @version 1.0
- * @date 2021/7/21 11:58
+ * @date 2021/7/21 11:46
  */
-public class TupApiClientApp {
+public class AntiSpoofClientApp {
 
     private static final String appId = PropertiesConfig.getAppId();
     private static final String apiKey = PropertiesConfig.getApiKey();
+    private static final String apiSecret = PropertiesConfig.getApiSecret();
 
     private static String filePath = "xxxxxxxx";
     private static String resourcePath;
 
     static {
         try {
-            resourcePath = TupApiClientApp.class.getResource("/").toURI().getPath();
+            resourcePath = AntiSpoofClientApp.class.getResource("/").toURI().getPath();
         } catch (URISyntaxException e) {
             e.printStackTrace();
         }
     }
 
-
     public static void main(String[] args) throws Exception {
-        TupApiClient client = new TupApiClient
-                // 年龄      TupApiEnum.AGE
-                // 性别      TupApiEnum.SEX
-                // 表情      TupApiEnum.EXPRESSION
-                // 颜值      TupApiEnum.FACE_SCORE
-                .Builder(appId, apiKey, TupApiEnum.AGE)
+        // 静默活体检测
+        AntiSpoofClient client = new AntiSpoofClient
+                .Builder(appId, apiKey, apiSecret)
                 .build();
         InputStream inputStream = new FileInputStream(new File(resourcePath + filePath));
         byte[] bytes = IoUtil.readBytes(inputStream);
+        String imageBase64 = Base64.getEncoder().encodeToString(bytes);
         System.out.println("请求地址：" + client.getHostUrl());
-        System.out.println(client.recognition("测试", bytes));
+        System.out.println(client.faceContrast(imageBase64, "jpg"));
     }
 }
