+++ src/main/java/cn/xfyun/demo/face/FaceStatusClientApp.java
@@ -1,7 +1,6 @@
-package cn.xfyun.demo;
+package cn.xfyun.demo.face;
 
-import cn.xfyun.api.ImageWordClient;
-import cn.xfyun.config.ImageWordEnum;
+import cn.xfyun.api.FaceStatusClient;
 import cn.xfyun.config.PropertiesConfig;
 import cn.hutool.core.io.IoUtil;
 
@@ -14,38 +13,34 @@ import java.util.Base64;
 /**
  * @author mqgao
  * @version 1.0
- * @date 2021/7/21 11:36
+ * @date 2021/7/21 11:54
  */
-public class ImageWordClientApp {
+public class FaceStatusClientApp {
 
     private static final String appId = PropertiesConfig.getAppId();
     private static final String apiKey = PropertiesConfig.getApiKey();
     private static final String apiSecret = PropertiesConfig.getApiSecret();
 
-    private static String filePath = "image/print.jpg";
+    private static String filePath = "xxxxxxxx";
     private static String resourcePath;
 
     static {
         try {
-            resourcePath = ImageWordClientApp.class.getResource("/").toURI().getPath();
+            resourcePath = FaceStatusClientApp.class.getResource("/").toURI().getPath();
         } catch (URISyntaxException e) {
             e.printStackTrace();
         }
     }
 
     public static void main(String[] args) throws Exception {
-        ImageWordClient client = new ImageWordClient
-                // 身份证识别      ImageWordEnum.IDCARD
-                // 营业执照识别    ImageWordEnum.BUSINESS_LICENSE
-                // 出租车发票识别  ImageWordEnum.TAXI_INVOICE
-                // 火车票识别      ImageWordEnum.TRAIN_TICKET
-                // 增值税发票识别  ImageWordEnum.INVOICE
-                // 多语种文字识别  ImageWordEnum.PRINTED_WORD
-                .Builder(appId, apiKey, apiSecret, ImageWordEnum.PRINTED_WORD)
+        // 配合式活体检测
+        FaceStatusClient client = new FaceStatusClient
+                .Builder(appId, apiKey, apiSecret)
                 .build();
         InputStream inputStream = new FileInputStream(new File(resourcePath + filePath));
         byte[] bytes = IoUtil.readBytes(inputStream);
         String imageBase64 = Base64.getEncoder().encodeToString(bytes);
-        System.out.println(client.imageWord(imageBase64, "jpg"));
+        System.out.println("请求地址：" + client.getHostUrl());
+        System.out.println(client.faceContrast(imageBase64, "jpg"));
     }
 }
