+++ src/main/java/cn/xfyun/demo/face/FaceVerificationClientApp.java
@@ -1,4 +1,4 @@
-package cn.xfyun.demo;
+package cn.xfyun.demo.face;
 
 import cn.xfyun.api.FaceVerificationClient;
 import cn.xfyun.config.PropertiesConfig;
