+++ src/main/java/cn/xfyun/demo/face/WatermarkVerificationClientApp.java
@@ -1,4 +1,4 @@
-package cn.xfyun.demo;
+package cn.xfyun.demo.face;
 
 import cn.xfyun.api.WatermarkVerificationClient;
 import cn.xfyun.config.PropertiesConfig;
