+++ src/main/java/cn/xfyun/demo/ocr/BankcardClientApp.java
@@ -1,49 +1,47 @@
-package cn.xfyun.demo;
+package cn.xfyun.demo.ocr;
 
-import cn.xfyun.api.ItrClient;
-import cn.xfyun.config.ItrEntEnum;
+import cn.xfyun.api.BankcardClient;
 import cn.xfyun.config.PropertiesConfig;
 import cn.hutool.core.io.IoUtil;
-
 import java.io.File;
 import java.io.FileInputStream;
+import java.io.IOException;
 import java.io.InputStream;
 import java.net.URISyntaxException;
 import java.util.Base64;
 
 /**
+ *     银行卡识别
+ *
  * @author mqgao
  * @version 1.0
- * @date 2021/7/21 11:41
+ * @date 2021/7/21 11:27
  */
-public class ItrClientApp {
+public class BankcardClientApp {
 
     private static final String appId = PropertiesConfig.getAppId();
     private static final String apiKey = PropertiesConfig.getApiKey();
-    private static final String apiSecret = PropertiesConfig.getApiSecret();
 
-    private static String filePath = "xxxxxxxx";
+    private static String filePath = "xxxxxxx";
     private static String resourcePath;
 
     static {
         try {
-            resourcePath = ItrClientApp.class.getResource("/").toURI().getPath();
+            resourcePath = BankcardClientApp.class.getResource("/").toURI().getPath();
         } catch (URISyntaxException e) {
             e.printStackTrace();
         }
     }
 
-    public static void main(String[] args) throws Exception {
-        // 拍照速算识别 公式识别
-        ItrClient client = new ItrClient
-                // 拍照速算识别  ItrEntEnum.MATH_ARITH
-                // 公式识别     ItrEntEnum.TEACH_PHOTO_PRINT
-                .Builder(appId, apiKey, apiSecret, ItrEntEnum.MATH_ARITH)
+    public static void main(String[] args) throws IOException {
+        // 银行卡识别
+        BankcardClient client = new BankcardClient
+                .Builder(appId, apiKey)
                 .build();
         InputStream inputStream = new FileInputStream(new File(resourcePath + filePath));
         byte[] bytes = IoUtil.readBytes(inputStream);
         String imageBase64 = Base64.getEncoder().encodeToString(bytes);
         System.out.println("请求地址：" + client.getHostUrl());
-        System.out.println(client.itr(imageBase64));
+        System.out.println(client.bankcard(imageBase64));
     }
 }
