+++ src/main/java/cn/xfyun/demo/ocr/FingerOcrClientApp.java
@@ -1,6 +1,6 @@
-package cn.xfyun.demo;
+package cn.xfyun.demo.ocr;
 
-import cn.xfyun.api.FaceCompareClient;
+import cn.xfyun.api.FingerOcrClient;
 import cn.xfyun.config.PropertiesConfig;
 import cn.hutool.core.io.IoUtil;
 import java.io.File;
@@ -12,39 +12,34 @@ import java.util.Base64;
 /**
  * @author mqgao
  * @version 1.0
- * @date 2021/7/21 11:50
+ * @date 2021/7/21 11:31
  */
-public class FaceCompareClientApp {
+public class FingerOcrClientApp {
 
     private static final String appId = PropertiesConfig.getAppId();
     private static final String apiKey = PropertiesConfig.getApiKey();
     private static final String apiSecret = PropertiesConfig.getApiSecret();
 
-    private static String filePath1 = "xxxxxxxx";
-    private static String filePath2 = "xxxxxxxx";
+    private static String filePath = "xxxxxxxx";
     private static String resourcePath;
 
     static {
         try {
-            resourcePath = FaceCompareClientApp.class.getResource("/").toURI().getPath();
+            resourcePath = FingerOcrClientApp.class.getResource("/").toURI().getPath();
         } catch (URISyntaxException e) {
             e.printStackTrace();
         }
     }
 
     public static void main(String[] args) throws Exception {
-        // 人脸比对
-        FaceCompareClient client = new FaceCompareClient
+        // 指尖文字识别
+        FingerOcrClient client = new FingerOcrClient
                 .Builder(appId, apiKey, apiSecret)
                 .build();
-        InputStream inputStream = new FileInputStream(new File(resourcePath + filePath1));
+        InputStream inputStream = new FileInputStream(new File(resourcePath + filePath));
         byte[] bytes = IoUtil.readBytes(inputStream);
-        String imageBase641 = Base64.getEncoder().encodeToString(bytes);
-
-        InputStream inputStream1 = new FileInputStream(new File(resourcePath + filePath2));
-        byte[] bytes1 = IoUtil.readBytes(inputStream1);
-        String imageBase642 = Base64.getEncoder().encodeToString(bytes1);
+        String imageBase64 = Base64.getEncoder().encodeToString(bytes);
         System.out.println("请求地址：" + client.getHostUrl());
-        System.out.println(client.faceCompare(imageBase641, "jpg", imageBase642, "jpg"));
+        System.out.println(client.fingerOcr(imageBase64));
     }
 }
