+++ src/main/java/cn/xfyun/demo/ocr/ItrClientApp.java
@@ -1,6 +1,7 @@
-package cn.xfyun.demo;
+package cn.xfyun.demo.ocr;
 
-import cn.xfyun.api.SilentDetectionClient;
+import cn.xfyun.api.ItrClient;
+import cn.xfyun.config.ItrEntEnum;
 import cn.xfyun.config.PropertiesConfig;
 import cn.hutool.core.io.IoUtil;
 
@@ -13,33 +14,36 @@ import java.util.Base64;
 /**
  * @author mqgao
  * @version 1.0
- * @date 2021/7/21 11:57
+ * @date 2021/7/21 11:41
  */
-public class SilentDetectionClientApp {
+public class ItrClientApp {
 
     private static final String appId = PropertiesConfig.getAppId();
     private static final String apiKey = PropertiesConfig.getApiKey();
+    private static final String apiSecret = PropertiesConfig.getApiSecret();
 
     private static String filePath = "xxxxxxxx";
     private static String resourcePath;
 
     static {
         try {
-            resourcePath = SilentDetectionClientApp.class.getResource("/").toURI().getPath();
+            resourcePath = ItrClientApp.class.getResource("/").toURI().getPath();
         } catch (URISyntaxException e) {
             e.printStackTrace();
         }
     }
 
     public static void main(String[] args) throws Exception {
-        // 活体检查sensetime
-        SilentDetectionClient client = new SilentDetectionClient
-                .Builder(appId, apiKey)
+        // 拍照速算识别 公式识别
+        ItrClient client = new ItrClient
+                // 拍照速算识别  ItrEntEnum.MATH_ARITH
+                // 公式识别     ItrEntEnum.TEACH_PHOTO_PRINT
+                .Builder(appId, apiKey, apiSecret, ItrEntEnum.MATH_ARITH)
                 .build();
         InputStream inputStream = new FileInputStream(new File(resourcePath + filePath));
         byte[] bytes = IoUtil.readBytes(inputStream);
-        String audioBase64 = Base64.getEncoder().encodeToString(bytes);
+        String imageBase64 = Base64.getEncoder().encodeToString(bytes);
         System.out.println("请求地址：" + client.getHostUrl());
-        System.out.println(client.silentDetection(audioBase64));
+        System.out.println(client.itr(imageBase64));
     }
 }
