+++ src/main/java/cn/xfyun/demo/ocr/PlaceRecClientApp.java
@@ -1,20 +1,16 @@
-package cn.xfyun.demo;
+package cn.xfyun.demo.ocr;
 
-import cn.hutool.core.io.IoUtil;
-import cn.xfyun.api.AntiSpoofClient;
+import cn.xfyun.api.PlaceRecClient;
 import cn.xfyun.config.PropertiesConfig;
+import cn.hutool.core.io.IoUtil;
+
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.InputStream;
 import java.net.URISyntaxException;
 import java.util.Base64;
 
-/**
- * @author mqgao
- * @version 1.0
- * @date 2021/7/21 11:46
- */
-public class AntiSpoofClientApp {
+public class PlaceRecClientApp {
 
     private static final String appId = PropertiesConfig.getAppId();
     private static final String apiKey = PropertiesConfig.getApiKey();
@@ -25,21 +21,20 @@ public class AntiSpoofClientApp {
 
     static {
         try {
-            resourcePath = AntiSpoofClientApp.class.getResource("/").toURI().getPath();
+            resourcePath = PlaceRecClientApp.class.getResource("/").toURI().getPath();
         } catch (URISyntaxException e) {
             e.printStackTrace();
         }
     }
 
     public static void main(String[] args) throws Exception {
-        // 静默活体检测
-        AntiSpoofClient client = new AntiSpoofClient
+        // 场所识别
+        PlaceRecClient client = new PlaceRecClient
                 .Builder(appId, apiKey, apiSecret)
                 .build();
         InputStream inputStream = new FileInputStream(new File(resourcePath + filePath));
-        byte[] bytes = IoUtil.readBytes(inputStream);
-        String imageBase64 = Base64.getEncoder().encodeToString(bytes);
-        System.out.println("请求地址：" + client.getHostUrl());
-        System.out.println(client.faceContrast(imageBase64, "jpg"));
+        byte[] imageByteArray = IoUtil.readBytes(inputStream);
+        String imageBase64 = Base64.getEncoder().encodeToString(imageByteArray);
+        System.out.println(client.send(imageBase64, "jpg"));
     }
 }
