+++ src/main/java/cn/xfyun/demo/speech/QbhClientApp.java
@@ -1,6 +1,6 @@
-package cn.xfyun.demo;
+package cn.xfyun.demo.speech;
 
-import cn.xfyun.api.BankcardClient;
+import cn.xfyun.api.QbhClient;
 import cn.xfyun.config.PropertiesConfig;
 import cn.hutool.core.io.IoUtil;
 import java.io.File;
@@ -8,40 +8,37 @@ import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.net.URISyntaxException;
-import java.util.Base64;
 
 /**
- *     银行卡识别
- *
- * @author mqgao
- * @version 1.0
- * @date 2021/7/21 11:27
- */
-public class BankcardClientApp {
+ * @author: <flhong2@iflytek.com>
+ * @description: 歌曲识别
+ * @version: v1.0
+ * @create: 2021-06-11 10:08
+ **/
+public class QbhClientApp {
 
     private static final String appId = PropertiesConfig.getAppId();
     private static final String apiKey = PropertiesConfig.getApiKey();
 
-    private static String filePath = "xxxxxxx";
+    private static String filePath = "audio/audio_qbh.wav";
     private static String resourcePath;
 
     static {
         try {
-            resourcePath = BankcardClientApp.class.getResource("/").toURI().getPath();
+            resourcePath = QbhClientApp.class.getResource("/").toURI().getPath();
         } catch (URISyntaxException e) {
             e.printStackTrace();
         }
     }
 
     public static void main(String[] args) throws IOException {
-        // 银行卡识别
-        BankcardClient client = new BankcardClient
-                .Builder(appId, apiKey)
+        // 歌曲识别
+        QbhClient client = new QbhClient.Builder(appId, apiKey)
                 .build();
         InputStream inputStream = new FileInputStream(new File(resourcePath + filePath));
         byte[] bytes = IoUtil.readBytes(inputStream);
-        String imageBase64 = Base64.getEncoder().encodeToString(bytes);
+        String result = client.send(bytes);
         System.out.println("请求地址：" + client.getHostUrl());
-        System.out.println(client.bankcard(imageBase64));
+        System.out.println("返回结果: " + result);
     }
 }
